The PPDs with missing converters, I think they are not needed as this kyocera style printers are able of printing a postscript compatible language I moved to maybe_broken. So if you do not want to use them, it is ok.
So I went on and searched for UTAX PPDs and found new PPDs which
do not need a filter. Also I changed the structure of the PPDs so they
are all sorted unter EU vs. Global / Language/model.ppd
So the TA350ci is sorted unter PPD/Utax/EU/English/TA350ci.ppd for instance.